python-image-gallery/requirements.txt
flask
uwsgi
psycopg2-binary
boto3

